Clear separation between Table modules and Repositories. Reduced dependencies and deleted unused queries and functions. Renamed table modules.
