+++ conduit-persistence/src/Conduit/Persistence/FavoritesTable.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE Arrows #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
@@ -13,91 +12,97 @@
 -- License     :  Apache License 2.0
 -- Maintainer  :  real-world-study-group@ugsmail.mailworks.org
 -- Stability   :  unstable
--- Lang. Ext.  :  Arrows - Required by Opaleye
---             :  FlexibleInstances - Required by Opaleye
---             :  MultiParamTypeClasses - Required by Opaleye
+-- Lang. Ext.  :  FlexibleInstances - For Opaleye table types
+--             :  MultiParamTypeClasses - For Opaleye table types
 --             :  TemplateHaskell - Lets Opaleye generate the mapping function
 --             :  NoImplicitPrelude - Use RIO instead
 --             :  GeneralizedNewtypeDeriving - Simplify newtype usage
 --             :  OverloadedStrings - Use Text literals
 --
--- Database interface for the "follows" relation, using the Opaleye mapper and
+-- Database interface for the "favorites" relation, using the Opaleye mapper and
 -- typesafe query and data manipulation DSL.
 -- See https://github.com/tomjaguarpaw/haskell-opaleye and the (outdated)
 -- tutorial here: https://www.haskelltutorials.com/opaleye/index.html
-module Conduit.Persistence.Follows
-  ( Follows
-  , findAllFollows
+module Conduit.Persistence.FavoritesTable
+  ( Favorite
+  , allFavoritesQ
   )
 where
 
-import qualified Control.Arrow                  ( )
 import qualified Data.Profunctor.Product        ( )
 import           Data.Profunctor.Product.TH     ( makeAdaptorAndInstance )
-import qualified Database.PostgreSQL.Simple    as PGS
 import qualified Opaleye                       as OE
+import qualified Conduit.Persistence.ArticlesTable
+                                               as PA
 import           Conduit.Persistence.DbConfig   ( schemaName )
-import qualified Conduit.Persistence.Users     as PU
+import qualified Conduit.Persistence.UsersTable
+                                               as PU
 import           RIO
 
 --------------------
 -- Table Setup
 --------------------
 
--- | Polymorphic type for the "follows" table.
-data FollowsT key
-  = Follows
-      { followerFk :: !key,
-        followeeFk :: !key
+-- | Polymorphic type for the "favorites" table.
+data FavoriteT userKey articleKey
+  = Favorite
+      { readerFk :: !userKey,
+        articleFk :: !articleKey
       }
   deriving (Show)
 
--- | Record that Opaleye uses to write to the "follows" table.
-type FollowsW = FollowsT PU.UserIdField -- user FK
+-- | Record that Opaleye uses to write to the "favorites" table.
+type FavoriteW
+  = FavoriteT PU.UserIdField -- reader FK
+                             PA.ArticleIdField -- favorite FK
 
--- | Record that Opaleye reads from the "follows" table.
-type FollowsR = FollowsW
+-- | Record that Opaleye reads from the "favorites" table.
+type FavoriteR = FavoriteW
 
--- | Typesafe Haskell record to interface with the application.
-type Follows = FollowsT PU.UserId -- user fk
+-- | Typesafe Haskell record to interface with the application. Under the hood,
+-- Opaleye converts between this application record and the above PostgreSQL
+-- read and write records.
+type Favorite
+  = FavoriteT PU.UserId -- reader
+                        PA.ArticleId -- favorite
 
-instance Display Follows where
+instance Display Favorite where
   display = displayShow
 
 -- | Template Haskell helper to create the mapping function between PostgreSQL
 -- records and the Haskell record used below.
-$(makeAdaptorAndInstance "pFollows" ''FollowsT)
+$(makeAdaptorAndInstance "pFavorite" ''FavoriteT)
 
 -- | The actual mapping setup tells Opaleye exactly how to map between the
 -- PostgreSQL records and the Haskell record. For each record, the function
 -- specifies the name of the table column and the constraints.
-followsTable :: OE.Table FollowsW FollowsR
-followsTable = OE.tableWithSchema
+favoritesTable :: OE.Table FavoriteW FavoriteR
+favoritesTable = OE.tableWithSchema
   schemaName
-  "follows"
-  (pFollows Follows
-    { followerFk = PU.pUserId (PU.UserId (OE.tableField "follower_fk"))
-    , followeeFk = PU.pUserId (PU.UserId (OE.tableField "followee_fk"))
+  "favorites"
+  (pFavorite Favorite
+    { readerFk  = PU.pUserId (PU.UserId (OE.tableField "reader_fk"))
+    , articleFk = PA.pArticleId (PA.ArticleId (OE.tableField "favorite_fk"))
     }
   )
 
 --------------------
--- Queries
+-- Basic Query
 --------------------
 
 -- | Retrieve all follow relations.
-selectFollows :: OE.Select FollowsR
-selectFollows = OE.selectTable followsTable
+allFavoritesQ :: OE.Select FavoriteR
+allFavoritesQ = OE.selectTable favoritesTable
 
 --------------------
 -- DB Access
 --------------------
 
--- | Find all following relations stored in the DB and return them.
+-- | Find all favorites stored in the DB and return them.
 -- Naming convention: DB retrievals are called "find".
-findAllFollows :: PGS.ConnectInfo -> IO [Follows]
-findAllFollows connInfo = do
-  conn   <- PGS.connect connInfo
-  result <- OE.runSelect conn selectFollows
-  PGS.close conn
-  return result
+-- findAllFavorites :: PGS.ConnectInfo -> IO [Favorite]
+-- findAllFavorites connInfo = do
+--   conn   <- PGS.connect connInfo
+--   result <- OE.runSelect conn allFavoritesQ
+--   PGS.close conn
+--   return result
