+++ conduit-persistence/src/Conduit/Persistence/UserRepository.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE Arrows #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE NoImplicitPrelude #-}
@@ -9,28 +10,34 @@
 -- License     :  Apache License 2.0
 -- Maintainer  :  real-world-study-group@ugsmail.mailworks.org
 -- Stability   :  unstable
--- Lang. Ext.  :  FlexibleInstances - Required by Opaleye
+-- Lang. Ext.  :  Arrows - For Opaleye queries
+--             :  FlexibleInstances - For Opaleye table types
 --             :  NoImplicitPrelude - Use RIO instead
 --             :  OverloadedStrings - Use Text literals
 --
 -- Abstract collection-like interface to the underlying persistence layer.
 module Conduit.Persistence.UserRepository
   ( readUser
-  , toDomain
+  , findAllUsers
   )
 where
 
 import           Control.Arrow                  ( (<<<)
                                                 , arr
+                                                , returnA
                                                 )
 import qualified Database.PostgreSQL.Simple    as PGS
 import qualified Opaleye                       as OE
+import           Opaleye                        ( (.===) )
 import qualified Conduit.Domain.User           as DU
 import qualified Conduit.Domain.Username       as DUN
 import qualified Conduit.Persistence.DbConfig  as DBC
-import qualified Conduit.Persistence.Users     as PU
+import qualified Conduit.Persistence.UsersTable
+                                               as PU
+import           Conduit.Persistence.PersistenceUtils
 import           RIO
 import           RIO.List                      as L
+
 
 -- | Returns a specific user stored in the underlying persistence mechanism
 -- Naming convention: Read repository operations are called "read".
@@ -41,15 +48,15 @@ readUser (DUN.Username uName) = do
   pUser    <- liftIO $ findUser connInfo uName
   case pUser of
     Nothing -> return $ Left ("Could not find user with username " <> uName)
-    Just u  -> return $ toDomain u
+    Just u  -> return $ toUser u
 
 -- | Helper function that converts the data obtained from the DB into a
 -- corresponding domain object. Because the domain types are shielded through
 -- smart constructors, we need to handle the case here that a user read from
 -- the DB does not conform to the domain restrictions. For now, we do it in the
 -- simplest possible way: return an error message.
-toDomain :: PU.User -> Either Text DU.User
-toDomain pu = case maybeUser of
+toUser :: PU.User -> Either Text DU.User
+toUser pu = case maybeUser of
   Just user -> Right user
   Nothing ->
     Left
@@ -68,7 +75,7 @@ toDomain pu = case maybeUser of
 -- DB Access
 --------------------
 -- Functions in the IO Monad that perform the actual database access, given a
--- connection string. These functions use Opaleye primitive that perform the
+-- connection string. These functions use Opaleye primitives that perform the
 -- mapping between Haskell records and Opaleye PostgreSQL records.
 
 -- | Find all users stored in the DB and return them.
@@ -86,6 +93,21 @@ findUser connInfo uName = do
   conn   <- PGS.connect connInfo
   result <- OE.runSelect
     conn
-    (PU.userByNameQ <<< arr (const $ OE.sqlStrictText uName)) -- Convert a parameter into an arrow via the const function.
+    (userByNameQ <<< arr (const $ OE.sqlStrictText uName)) -- Convert a parameter into an arrow via the const function.
   PGS.close conn
   return $ L.headMaybe result
+
+--------------------
+-- Compund Queries
+--------------------
+-- The queries below are written in the typesafe Opaleye query DSL. Opaleye
+-- translates them into actual SQL statements that can be executed against the
+-- DBMS. The query statements specified below are similar to prepared
+-- statements; they need to be executed separately.
+-- Queries return Opaleye PostgreSQL "Read" records.
+
+userByNameQ :: OE.SelectArr (F OE.SqlText) PU.UserR
+userByNameQ = proc uName -> do
+  row <- PU.allUsersQ -< ()
+  OE.restrict -< PU.userUsername row .=== uName
+  returnA -< row
